+++ docker-compose.yaml
@@ -3,7 +3,7 @@ version: "3.8"
 services:
   frontend:
     build:
-      context: ./fe2
+      context: ./fe
       dockerfile: Dockerfile
     container_name: IOD-frontend
     ports:
