+++ fe/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-masterclass",
+  "name": "IOD-FrontEnd",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
