+++ fe/src/App.js
@@ -28,15 +28,15 @@ function App() {
   // 감정에 따른 색상을 반환하는 함수
   const getEmotionColor = (x, y) => {
     if (x > 0 && y >= 0) {
-      return "bg-yellow-200";
+      return "yellow";
     } else if (x < 0 && y > 0) {
-      return "bg-red-200";
+      return "red";
     } else if (x < 0 && y <= 0) {
-      return "bg-blue-200";
+      return "blue";
     } else if (x > 0 && y < 0) {
-      return "bg-green-200";
+      return "green";
     }
-    return "bg-gray-200"; // 기본 색상
+    return "gray"; // 기본 색상
   };
 
   const [token, setToken] = useState("");
@@ -45,9 +45,12 @@ function App() {
   const [pliKey, setPliKey] = useState("");
   const [pliName, setPliName] = useState("");
   const [playlists, setPlaylists] = useState([]);
+  const [track, setTrack] = useState(null);
   const [isPlaying, setIsPlaying] = useState(false);
   const [userInfo, setUserInfo] = useState(null);
-  const [emotion, setEmotion] = useState("");
+  const [emotion, setEmotion] = useState("teal");
+  const [currentTime, setCurrentTime] = useState(0);
+  const [duration, setDuration] = useState(0);
   // 유저 토큰 받아오기
   // navigate 정의
 
@@ -82,7 +85,7 @@ function App() {
 
         const userData = userResponse.data;
         setUserInfo(userData);
-        // console.log(userData);
+        console.log(userData);
 
         // 이메일을 기반으로 백엔드에 해당 사용자가 있는지 확인
         const checkUserResponse = await axios.get(
@@ -126,7 +129,7 @@ function App() {
         console.log("Spotify Web Playback SDK is ready");
         if (token) {
           const playerInstance = new window.Spotify.Player({
-            name: "Web Playback SDK Player",
+            name: "InsideOutDJ Webplayer",
             getOAuthToken: (cb) => {
               cb(token);
             },
@@ -145,7 +148,22 @@ function App() {
 
           playerInstance.addListener("player_state_changed", (state) => {
             console.log(state);
+            if (!state) {
+              return;
+            }
+
             setIsPlaying(!state.paused);
+            setCurrentTime(state.position / 1000);
+            setDuration(state.duration / 1000);
+
+            const currentTrack = state.track_window.current_track;
+            setTrack({
+              name: currentTrack.name,
+              artists: currentTrack.artists
+                .map((artist) => artist.name)
+                .join(", "),
+              album: currentTrack.album.images.map((image) => image.url), // Get the cover image URL
+            });
           });
 
           playerInstance.connect().then((success) => {
@@ -339,6 +357,19 @@ function App() {
     emotion,
     setEmotion,
     pliKey,
+    track: track,
+    seekTo: (progress) => {
+      const newTime = progress;
+      player.seek(newTime * 1000).then(() => {
+        console.log(`재생시간 ${newTime}로 설정되었습니다.`);
+      });
+      setCurrentTime(newTime);
+    },
+    userInfo: userInfo,
+    currentTime: currentTime,
+    duration: duration,
+    setPliName,
+    setEmotion,
   };
   return (
     <AppRoutes
