+++ fe/src/assets/empty_album.png
