+++ fe/src/components/MenuSlide.js
