+++ fe/src/components/PlaylistsPage.js
@@ -81,15 +81,27 @@ function PlaylistsPage({
 
   const getEmotionColor = (x, y) => {
     if (x > 0 && y >= 0) {
-      return "bg-yellow-200";
+      return "yellow";
     } else if (x < 0 && y > 0) {
-      return "bg-red-200";
+      return "red";
     } else if (x < 0 && y <= 0) {
-      return "bg-blue-200";
+      return "blue";
     } else if (x > 0 && y < 0) {
-      return "bg-green-200";
+      return "green";
     }
-    return "bg-gray-200"; // 기본 색상
+    return "gray"; // 기본 색상
+  };
+  const getEmotionColor2 = (x, y) => {
+    if (x > 0 && y >= 0) {
+      return "bg-yellow-100";
+    } else if (x < 0 && y > 0) {
+      return "bg-red-100";
+    } else if (x < 0 && y <= 0) {
+      return "bg-blue-100";
+    } else if (x > 0 && y < 0) {
+      return "bg-green-100";
+    }
+    return "bg-gray-100"; // 기본 색상
   };
   const goToDiary = () => {
     navigate("/diary");
@@ -102,11 +114,11 @@ function PlaylistsPage({
       <PlaylistList>
         {playlists.map((playlist) => (
           <PlaylistItem
-            className={getEmotionColor(
+            className={getEmotionColor2(
               playlist.emotion_analysis.normalized_emotion.x,
               playlist.emotion_analysis.normalized_emotion.y
             )}
-            key={playlist.id}
+            key={playlist.playlist_id}
             onClick={() =>
               handlePlaylistClick(
                 playlist.playlist_id,
